Create README.md
# iCoder-Ibraheem-Awan

Welcome to the iCoder project! This project includes a set of web pages for showcasing different features.

## Files
- **Index.html**: Main page
- **About.html**: About the project
- **Contact.html**: Contact form page

## Installation
To run this project, simply open the HTML files in your web browser.

## Contributing
Contributions are welcome! Please fork this repository and submit a pull request.

## License
This project is licensed under the MIT License.